requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
boto3==1.34.144
python-dotenv==1.0.0
Flask-Mail==0.9.1
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7